Use External Editor for Images (#453)
Developing the internal Image Editor is a lot of trouble not worth the effort when we already have better programs suited to the task. Notably, even MS Paint supports more features than the current Image Editor. So, I'm dropping work on the internal image editor.
Also, removed the Image crop window as that too is supported by other image editors.
The external editor can be set from the screenshot settings page and is set by default to MS Paint.